+++ src/crescent.rs
@@ -1,4 +1,4 @@
-use anyhow::{Context, Result};
+use anyhow::{anyhow, Context, Result};
 use std::{env, fs, path::PathBuf};
 
 pub fn crescent_dir() -> Result<PathBuf> {
@@ -15,11 +15,21 @@ pub fn crescent_dir() -> Result<PathBuf> {
     Ok(crescent_dir)
 }
 
-pub fn get_profile_path(profile_name: String) -> Result<PathBuf> {
-    let mut crescent_dir = crescent_dir()?;
-    crescent_dir.push("profiles");
-    crescent_dir.push(profile_name + ".toml");
-    Ok(crescent_dir)
+pub fn get_profile_path(profile: String) -> Result<PathBuf> {
+    match fs::canonicalize(&profile) {
+        Ok(path) => Ok(path),
+        Err(_) => {
+            let mut crescent_dir = crescent_dir()?;
+            crescent_dir.push("profiles");
+            crescent_dir.push(profile + ".toml");
+
+            if crescent_dir.exists() && crescent_dir.is_file() {
+                return Ok(crescent_dir);
+            }
+
+            Err(anyhow!("Profile not found."))
+        }
+    }
 }
 
 #[cfg(test)]
@@ -34,6 +44,23 @@ mod tests {
         home_path.push(".crescent");
 
         assert_eq!(crescent_dir()?, home_path);
+
+        Ok(())
+    }
+
+    #[test]
+    fn unit_get_profile_path() -> Result<()> {
+        let mut path = get_profile_path(String::from("example"))?;
+        assert!(path.exists());
+        assert!(path.is_file());
+
+        path = get_profile_path(String::from("./profiles/example.toml"))?;
+        assert!(path.exists());
+        assert!(path.is_file());
+
+        let err = get_profile_path(String::from("profile/does/not/exist")).unwrap_err();
+
+        assert_eq!(format!("{}", err), "Profile not found.");
 
         Ok(())
     }