+++ tests/util.rs
@@ -21,8 +21,51 @@ pub fn delete_app_folder(name: &str) -> Result<()> {
     Ok(())
 }
 
+pub fn shutdown_long_running_service(name: &str) -> Result<()> {
+    let mut cmd = get_base_command();
+
+    cmd.args(["signal", name, "15"]);
+
+    cmd.assert()
+        .success()
+        .stdout(predicate::str::contains("Signal sent."));
+
+    // Sleeping to make sure the process exited
+    thread::sleep(std::time::Duration::from_secs(1));
+
+    Ok(())
+}
+
+// Solves issues with assert_cmd not finding the binary when using cross.
+// https://github.com/assert-rs/assert_cmd/issues/139#issuecomment-1200146157
+pub fn get_base_command() -> Command {
+    let mut cmd;
+    let path = assert_cmd::cargo::cargo_bin("cres");
+    if let Some(runner) = find_runner() {
+        let mut runner = runner.split_whitespace();
+        cmd = Command::new(runner.next().unwrap());
+        for arg in runner {
+            cmd.arg(arg);
+        }
+        cmd.arg(path);
+    } else {
+        cmd = Command::new(path);
+    }
+    cmd
+}
+
+// https://github.com/assert-rs/assert_cmd/issues/139#issuecomment-1200146157
+fn find_runner() -> Option<String> {
+    for (key, value) in std::env::vars() {
+        if key.starts_with("CARGO_TARGET_") && key.ends_with("_RUNNER") && !value.is_empty() {
+            return Some(value);
+        }
+    }
+    None
+}
+
 pub fn start_short_lived_command(name: &str) -> Result<()> {
-    let mut cmd = Command::cargo_bin("cres")?;
+    let mut cmd = get_base_command();
     cmd.args(["start", "/bin/echo", "-n", name]);
 
     cmd.assert()
@@ -36,7 +79,7 @@ pub fn start_short_lived_command(name: &str) -> Result<()> {
 }
 
 pub fn start_long_running_service(name: &str) -> Result<()> {
-    let mut cmd = Command::cargo_bin("cres")?;
+    let mut cmd = get_base_command();
 
     let args = [
         "start",
@@ -60,7 +103,7 @@ pub fn start_long_running_service(name: &str) -> Result<()> {
 }
 
 pub fn check_app_is_running(name: &str) -> Result<bool> {
-    let mut cmd = Command::cargo_bin("cres")?;
+    let mut cmd = get_base_command();
     cmd.args(["status", name]);
 
     let binding = cmd.assert().success();